+++ components/src/main/java/com/mercadolibre/android/andesui/list/utils/AndesListAdapter.kt
@@ -96,7 +96,6 @@ class AndesListAdapter(
                 is AndesListViewItemSimple -> bindSimpleItem(andesListItemConfig, dividerEnabled)
                 is AndesListViewItemChevron -> bindChevronItem(andesListItemConfig, dividerEnabled)
             }
-
             itemView.accessibilityDelegate = AndesListViewItemAccessibilityDelegate(andesListItemConfig)
         }